[flang][Lower] Move getHashValue and isEqual implementations to Utils.cpp
This is intended to reduce the memory usage while compiling flang sources.

There were over 7500 instantiations of function templates defined in the
Utils.h file. Most of them were not referenced anywhere outside, except
for specialized implementations of getHashValue and isEqual in
IterationSpace.cpp. These function were also moved to Utils.cpp.
